refactor moving popup enums and table.to_rows
